requirements.txt
streamlit
langchain
langchain-community
pypdf
python-dotenv
sentence-transformers
faiss-cpu
transformers
openai
